+++ iftttie/channels/clock.py
@@ -13,8 +13,8 @@ from iftttie.types_ import Event
 
 
 class Clock(BaseChannel):
-    def __init__(self, key: str, interval: timedelta, title: Optional[str] = None):
-        self.key = key
+    def __init__(self, channel_id: str, interval: timedelta, title: Optional[str] = None):
+        self.channel_id = channel_id
         self.interval = interval.total_seconds()
         self.title = title
 
@@ -22,7 +22,8 @@ class Clock(BaseChannel):
         for i in count(start=1):
             logger.trace('Sleeping for {interval} seconds…', interval=self.interval)
             await sleep(self.interval)
-            await context.trigger_event(Event(key=f'clock:{self.key}', value=i, title=self.title))
+            # FIXME: set unit.
+            await context.trigger_event(Event(channel_id=f'clock:{self.channel_id}', value=i, title=self.title))
 
     def __str__(self) -> str:
-        return f'Clock(key={self.key!r}, interval={self.interval!r}, title={self.title!r})'
+        return f'Clock(key={self.channel_id!r}, interval={self.interval!r}, title={self.title!r})'